feat: split routines from edit user data and change name from ShowUserData to UserDetails
