+++ src/pages/coach/pages/Users.tsx
@@ -181,61 +181,73 @@ const Users = () => {
 														</div>
 													</div>
 												</div>
-												<h3 className="text-2xl font-bold mt-10 mb-6 text-gray-800">Rutinas</h3>
-												<div className="space-y-3 flex flex-col gap-2">
-													{userRoutines.length > 0 ? (
-														userRoutines.map((routine) => (
-															<Link className="bg-gray-100 hover:bg-gray-200 text-gray-800 font-semibold py-3 px-6 rounded-lg shadow-md transition duration-300 ease-in-out" to={`/coach/routines/${routine.id}`} key={routine.id}>{routine.name}</Link>
-														))
+												<div className="px-8">
+													<div className="flex justify-between items-center h-content mt-10 mb-6">
+														<h3 className="text-2xl font-bold text-gray-800">Rutinas</h3>
+														<Link to={`/coach/users/${selectedUser?.id}/routines`} className="text-gray-700 font-semibold hover:text-gray-900 transition duration-300 ease-in-out">
+															<svg xmlns="http://www.w3.org/2000/svg" className="h-6 w-6 inline-block mr-2" viewBox="0 0 20 20" fill="currentColor">
+																<path d="M13.586 3.586a2 2 0 112.828 2.828l-.793.793-2.828-2.828.793-.793zM11.379 5.793L3 14.172V17h2.828l8.38-8.379-2.83-2.828z" />
+															</svg>
+															Editar
+														</Link>
+													</div>
+													<div className="space-y-3 flex flex-col gap-2">
+														{userRoutines.length > 0 ? (
+															userRoutines.map((routine) => (
+																<Link className="bg-gray-100 hover:bg-gray-200 text-gray-800 font-semibold py-3 px-6 rounded-lg shadow-md transition duration-300 ease-in-out" to={`/coach/routines/${routine.id}`} key={routine.id}>{routine.name}</Link>
+															))
+														) : (
+															<p className="text-gray-600 italic">El usuario no tiene rutinas asignadas.</p>
+														)}
+													</div>
+												</div>
+												<div className="px-8">
+													<div className="flex flex-col sm:flex-row justify-between items-center h-auto mt-10 mb-6">
+														<h3 className="text-2xl font-bold text-gray-800 mb-4 sm:mb-0">Últimas Medidas</h3>
+														<Link to={`/coach/users/${user.id}/measurements`}
+															className="px-6 py-3 bg-blue-600 text-white rounded-full hover:bg-blue-700 transition duration-300 ease-in-out shadow-lg"
+														>
+															Agregar Medida
+														</Link>
+													</div>
+													{userMeasurements.length > 0 ? (
+														<ul className="space-y-6">
+															{userMeasurements.map((measurement) => (
+																<li key={measurement.id} className="bg-gray-50 rounded-xl p-4 sm:p-6 shadow-md">
+																	<p className="text-lg font-semibold mb-3 text-gray-800">
+																		<strong>Fecha:</strong>{" "}
+																		{measurement.measurement_date
+																			? new Date(measurement.measurement_date).toLocaleDateString()
+																			: ""}
+																	</p>
+																	<div className="grid grid-cols-2 gap-4 mb-6">
+																		<span className="text-gray-700"><strong>Peso: </strong>{measurement?.weight ? measurement?.weight : 'NA'} kg</span>
+																		<span className="text-gray-700"><strong>Altura: </strong>{measurement?.height ? measurement?.height : 'NA'} cm</span>
+																	</div>
+
+																	<div className="relative w-full sm:w-64 h-96 mx-auto">
+																		{selectedUser?.gender === 'masculino' ? (
+																			<img src="/assets/men-human-outline.png" alt="Men Silhouette" className="w-full h-full flex justify-center items-center object-contain" />
+																		) : (
+																			<img src="/assets/woman-human-outline.png" alt="Woman Silhouette" className="w-full h-full flex justify-center items-center object-contain" />
+																		)}
+
+																		<span className="absolute top-24 left-4 bg-blue-100 px-2 py-1 rounded-full text-sm">{measurement.left_arm} cm</span>
+																		<span className="absolute top-24 right-4 bg-blue-100 px-2 py-1 rounded-full text-sm">{measurement.right_arm} cm</span>
+
+																		<span className="absolute top-32 left-1/2 transform -translate-x-1/2 bg-blue-100 px-2 py-1 rounded-full text-sm">{measurement.upper_waist} cm</span>
+																		<span className="absolute top-40 left-1/2 transform -translate-x-1/2 bg-blue-100 px-2 py-1 rounded-full text-sm">{measurement.lower_waist} cm</span>
+
+																		<span className="absolute bottom-32 left-8 bg-blue-100 px-2 py-1 rounded-full text-sm">{measurement.left_thigh} cm</span>
+																		<span className="absolute bottom-32 right-8 bg-blue-100 px-2 py-1 rounded-full text-sm">{measurement.right_thigh} cm</span>
+																	</div>
+																</li>
+															))}
+														</ul>
 													) : (
-														<p className="text-gray-600 italic">El usuario no tiene rutinas asignadas.</p>
+														<p className="text-gray-600 italic">El usuario no tiene medidas registradas.</p>
 													)}
 												</div>
-												<div className="flex flex-col sm:flex-row justify-between items-center h-auto mt-10 mb-6">
-													<h3 className="text-2xl font-bold text-gray-800 mb-4 sm:mb-0">Últimas Medidas</h3>
-													<Link to={`/coach/users/${user.id}/measurements`}
-														className="px-6 py-3 bg-blue-600 text-white rounded-full hover:bg-blue-700 transition duration-300 ease-in-out shadow-lg"
-													>
-														Agregar Medida
-													</Link>
-												</div>
-												{userMeasurements.length > 0 ? (
-													<ul className="space-y-6">
-														{userMeasurements.map((measurement) => (
-															<li key={measurement.id} className="bg-gray-50 rounded-xl p-4 sm:p-6 shadow-md">
-																<p className="text-lg font-semibold mb-3 text-gray-800">
-																	<strong>Fecha:</strong>{" "}
-																	{measurement.measurement_date
-																		? new Date(measurement.measurement_date).toLocaleDateString()
-																		: ""}
-																</p>
-																<div className="grid grid-cols-2 gap-4 mb-6">
-																	<span className="text-gray-700"><strong>Peso: </strong>{measurement?.weight ? measurement?.weight : 'NA'} kg</span>
-																	<span className="text-gray-700"><strong>Altura: </strong>{measurement?.height ? measurement?.height : 'NA'} cm</span>
-																</div>
-
-																<div className="relative w-full sm:w-64 h-96 mx-auto">
-																	{selectedUser?.gender === 'masculino' ? (
-																		<img src="/assets/men-human-outline.png" alt="Men Silhouette" className="w-full h-full flex justify-center items-center object-contain" />
-																	) : (
-																		<img src="/assets/woman-human-outline.png" alt="Woman Silhouette" className="w-full h-full flex justify-center items-center object-contain" />
-																	)}
-
-																	<span className="absolute top-24 left-4 bg-blue-100 px-2 py-1 rounded-full text-sm">{measurement.left_arm} cm</span>
-																	<span className="absolute top-24 right-4 bg-blue-100 px-2 py-1 rounded-full text-sm">{measurement.right_arm} cm</span>
-
-																	<span className="absolute top-32 left-1/2 transform -translate-x-1/2 bg-blue-100 px-2 py-1 rounded-full text-sm">{measurement.upper_waist} cm</span>
-																	<span className="absolute top-40 left-1/2 transform -translate-x-1/2 bg-blue-100 px-2 py-1 rounded-full text-sm">{measurement.lower_waist} cm</span>
-
-																	<span className="absolute bottom-32 left-8 bg-blue-100 px-2 py-1 rounded-full text-sm">{measurement.left_thigh} cm</span>
-																	<span className="absolute bottom-32 right-8 bg-blue-100 px-2 py-1 rounded-full text-sm">{measurement.right_thigh} cm</span>
-																</div>
-															</li>
-														))}
-													</ul>
-												) : (
-													<p className="text-gray-600 italic">El usuario no tiene medidas registradas.</p>
-												)}
 											</div>
 										</DialogDescription>
 									</DialogHeader>
